move CL calculations to own class
