+++ Dogfighter/Assets/Scenes/Nivel 0/fisica_local.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd5aa5113ec4e754482730157b5906b1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
