+++ Dogfighter/Assets/Scenes/Nivel 0/fisica_local/CL_AoA.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e278399b4dc29742aba103a7bfae1ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
